requirements.txt
streamlit
google-generativeai
pickle
numpy
pandas
scikit-learn
dotenv